requirements.txt
python-telegram-bot==11.1.0
environs==9.5.0
redis==4.3.4
